Update to controller to allow user to add category to task.

+++ controllers/taskController.js
@@ -1,19 +1,20 @@
 const Task = require("../models/taskModel");
 
 const getTasks = async (req, res) => {
-  const tasks = await Task.find();
+  const tasks = await Task.find({ user: req.user.id });
   res.json(tasks);
 };
 
 const setTask = async (req, res) => {
   try {
-    const { title, energyRequired, urgency, dueDate } = req.body;
+    const { title, energyRequired, urgency, dueDate, category } = req.body;
 
     const task = await Task.create({
       title,
       energyRequired,
       urgency,
       dueDate,
+      category,
       user: req.user.id,
     });
 
@@ -26,6 +27,11 @@ const setTask = async (req, res) => {
 const updateTask = async (req, res) => {
   const task = await Task.findById(req.params.id);
   if (!task) return res.status(404).json({ message: "Not found" });
+
+  if (task.user.toString() !== req.user.id) {
+    return res.status(401).json({ message: "User not authorised" });
+  }
+
   const updated = await Task.findByIdAndUpdate(req.params.id, req.body, {
     returnDocument: "after",
   });
@@ -36,6 +42,10 @@ const updateTask = async (req, res) => {
 const deleteTask = async (req, res) => {
   const task = await Task.findById(req.params.id);
   if (!task) return res.status(404).json({ message: "Not found" });
+
+  if (task.user.toString() !== req.user.id) {
+    return res.status(401).json({ message: "User not authorised" });
+  }
 
   await task.deleteOne();
   res.json({ id: req.params.id });
